refactor: decompose background service worker and fix React patterns
Split the 1025-line background/index.ts into focused modules:
- helpers.ts — dispatchAsync() utility, sendToPort()
- githubApi.ts — all GitHub REST calls and auth helpers
- llmService.ts — Anthropic streaming, summary generation
- skillResolver.ts — skill matching, fetching, caching
- index.ts — slim 130-line dispatcher with a handler registry

Other improvements:
- Extract parsePromptSuggestions() to shared/parsing.ts (was duplicated
  in background + ChatPanel)
- Replace nested ternaries with lookup maps (FILE_STATUS_LABELS,
  FILE_STATUS_INDICATOR, REVIEW_STATE_COLORS)
- Wrap inline callbacks in useCallback (handleSummaryLoading/Ready)
- Fix suppressed exhaustive-deps in ChatPanel and PRDashboard by using
  refs for callback props (updated via useEffect per linter rules)
- Use stable keys instead of array indices in ReviewQueue and ChatInput
- Replace magic string "2022-11-28" with GITHUB_API_VERSION constant
- Replace magic number 30 with MAX_COMMIT_DETAILS constant

Part of #21
